preparing to handle colis & envois in the backend
Took 40 minutes

+++ backend/src/main/java/com/moncolisapp/backend/dto/ColisDTO.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * DTO for {@link com.moncolisapp.backend.entities.Colis}
@@ -17,9 +16,9 @@ import java.math.BigDecimal;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ColisDTO implements Serializable {
     private EnvoisDTO idEnvoi;
-    private BigDecimal hauteur;
-    private BigDecimal largeur;
-    private BigDecimal longueur;
-    private BigDecimal poidsColis;
-    private BigDecimal volumeColis;
+    private double hauteur;
+    private double largeur;
+    private double longueur;
+    private double poidsColis;
+    private double volumeColis;
 }
